requirements.txt
beautifulsoup4>=4.9.3
black>=24.10.0
celery>=5.3.1
Django>=5.1.3
django-filter>=24.3
djangorestframework>=3.15.2
psycopg2>=2.9.10
python-dotenv>=1.0.1
redis>=5.0.0
requests>=2.25.1
